ADDED Deleting Polls ADDED Swal2 FIX Validation on CreatePollDialog

+++ src/PollTool.Server/Controllers/PollsController.cs
@@ -25,9 +25,9 @@ namespace PollTool.Server.Controllers
 
         private static List<ApiPoll> _debugPolls = new List<ApiPoll>
         {
-            new ApiPoll { Description  ="Tolle Umfrage 1", PollId = 1, Title = "Umfrage 1", DoneByUser = true},
-            new ApiPoll { Description  ="Tolle Umfrage 2", PollId = 2, Title = "Umfrage 2", DoneByUser = true} ,
-            new ApiPoll { Description  ="Tolle Umfrage 3", PollId = 3, Title = "Umfrage 3", DoneByUser = false} ,
+            new ApiPoll { Description  ="Tolle Umfrage 1", PollId = 1, Title = "Umfrage 1", DoneByUser = true, OwnedByUser = false},
+            new ApiPoll { Description  ="Tolle Umfrage 2", PollId = 2, Title = "Umfrage 2", DoneByUser = true, OwnedByUser = false},
+            new ApiPoll { Description  ="Tolle Umfrage 3", PollId = 3, Title = "Umfrage 3", DoneByUser = false, OwnedByUser = true} ,
         };
 
         public PollsController(PollContext context, ILogger<PollsController> logger)
@@ -53,13 +53,15 @@ namespace PollTool.Server.Controllers
             {
                 //var uid = HttpContext.Connection.Id
                 var uid = HttpContext.Connection.RemoteIpAddress.Address;
+                var ip = HttpContext.Connection.RemoteIpAddress.ToString();
                 var polls = await _context.Polls.Select(p => new ApiPoll
                 {
                     Title = p.Title,
                     Description = p.Description,
                     PollId = p.PollId,
                     DoneByUser = _context.Questions.Where(q => q.PollId == p.PollId)
-                    .Any(q => _context.UserAnswers.Any(ua => ua.QuestionId == q.QuestionId && ua.UserAnswerId == uid))
+                    .Any(q => _context.UserAnswers.Any(ua => ua.QuestionId == q.QuestionId && ua.UserAnswerId == uid)),
+                    OwnedByUser = p.CreatedBy == ip
                 }).ToListAsync();
                 response.Polls = polls;
                 response.Success = true;
@@ -81,9 +83,9 @@ namespace PollTool.Server.Controllers
         {
             if (!request.IsValid()) return BadRequest();
 
-            if(_debugPolls.Any(p => p.Title == request.Title)) return new CreatePollResponse { ErrorMessage = $"Umfrage '{request.Title}' existiert bereits!" };
+            if (_debugPolls.Any(p => p.Title == request.Title)) return new CreatePollResponse { ErrorMessage = $"Umfrage '{request.Title}' existiert bereits!" };
             var response = new CreatePollResponse();
-            _debugPolls.Add( new ApiPoll { Title = request.Title, Description = request.Description, DoneByUser = false });
+            _debugPolls.Add(new ApiPoll { Title = request.Title, Description = request.Description, DoneByUser = false, OwnedByUser = true, PollId = _debugPolls.Max(p => p.PollId) + 1});
             response.Success = true;
             return response;
 
@@ -113,17 +115,28 @@ namespace PollTool.Server.Controllers
             return response;
         }
 
-        // DELETE: api/Polls/5
+        // POST: api/Polls/DeletePoll
         [HttpPost]
         public async Task<ActionResult<BaseResponse>> DeletePoll([FromBody] DeletePollRequest request)
         {
             if (!request.IsValid()) return BadRequest();
-            //TODO: Log Exceptions
+
+            if (_debugPolls.FirstOrDefault(p => p.PollId == request.PollID) is ApiPoll pollToDelete)
+            {                
+                _debugPolls.Remove(pollToDelete);
+                return new BaseResponse { Success = true };
+
+            }
+            return new BaseResponse { ErrorMessage = "Poll not found!" };
+            var ip = HttpContext.Connection.RemoteIpAddress.ToString();
             if (await _context.Polls.FindAsync(request.PollID) is DbPoll foundPoll)
             {
+                if (foundPoll.CreatedBy != ip) return Unauthorized();
+
                 _context.Polls.Remove(foundPoll);
                 await _context.SaveChangesAsync();
                 return new BaseResponse { Success = true };
+
             }
             return new BaseResponse { ErrorMessage = "Poll not found!" };
